part 1

+++ libft/ft_memcpy.c
@@ -6,13 +6,13 @@
 /*   By: rzamolo- <rzamolo-@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 17:04:05 by rzamolo-          #+#    #+#             */
-/*   Updated: 2022/09/15 11:06:26 by rzamolo-         ###   ########.fr       */
+/*   Updated: 2022/09/16 12:55:20 by rzamolo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memcpy(void *dst, const void *src, unsigned int n)
+void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	unsigned int	position;
